+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
     - only `add_variable()` can expand the object
 - Added support for single value replacement in `VectorContainer` and
   `ModelInterface` (and, in turn, `BaseModel`).
+- Added initial version of `eval()` method to `VectorContainer` (and, in turn,
+  `BaseModel` and `BaseLinker`).
 - Added new `size` property to `VectorContainer` and `ModelInterface` (and, in
   turn, `BaseModel`), as well as `BaseLinker`, to report the number of elements
   in the objects' arrays.